Fix tests by adding renorm
I tried printing the index of the input that was tiled over and the thread aggregate before and after setting the max value, For some reason it had -381 as it's value after setting max value.
For some reason I could not find the function in the debugger which sets the max value and I could not find the value -381 in the debugger either. So this seemed like some random issue and I was going no where. Tried dumping the sass and went through the function but could not find anything useful there as well.
So I simply used the += operator to see if the update, deposit,renorm in that sequence worked. And to my surprise it did. So I just called renorm after adding the ITEMS_PER_THREAD number of elements even though we statically check if it is less than endurance.

Signed-off-by: Shreyas Atre <shreyasa@nvidia.com>